+++ core/src/main/java/io/snabble/sdk/payment/PaymentCredentials.java
@@ -40,6 +40,7 @@ import io.snabble.sdk.R;
 import io.snabble.sdk.Snabble;
 import io.snabble.sdk.payment.data.GiropayAuthorizationData;
 import io.snabble.sdk.payment.data.GiropayData;
+import io.snabble.sdk.payment.data.PayoneData;
 import io.snabble.sdk.payment.externalbilling.data.ExternalBillingPaymentCredentials;
 import io.snabble.sdk.payment.payone.sepa.PayoneSepaData;
 import io.snabble.sdk.utils.GsonHolder;
@@ -151,18 +152,6 @@ public class PaymentCredentials {
         private String alias;
         private String expiryMonth;
         private String expiryYear;
-    }
-
-    private static class PayoneData {
-        PayoneData(String pseudoCardPAN, String name, String userID) {
-            this.pseudoCardPAN = pseudoCardPAN;
-            this.name = name;
-            this.userID = userID;
-        }
-
-        private final String pseudoCardPAN;
-        private final String name;
-        private final String userID;
     }
 
     private static class TegutEmployeeCard {
@@ -470,18 +459,23 @@ public class PaymentCredentials {
     /**
      * Encrypts and stores a payone pseudo card pan.
      */
-    public static PaymentCredentials fromPayone(String pseudocardpan,
-                                                String truncatedcardpan,
-                                                PaymentCredentials.Brand brand,
-                                                String cardexpiredate,
-                                                String lastname,
-                                                String userId,
-                                                String projectId) {
-        if (pseudocardpan == null) {
-            return null;
-        }
-
-        PaymentCredentials pc = new PaymentCredentials();
+    @Nullable
+    public static PaymentCredentials fromPayone(
+            @NonNull final String pseudoCardPan,
+            @NonNull final String truncatedCardPan,
+            @NonNull final PaymentCredentials.Brand brand,
+            @NonNull final String cardExpiryDate,
+            @NonNull final String lastname,
+            @NonNull final String street,
+            @NonNull final String zip,
+            @NonNull final String city,
+            @NonNull final String country,
+            @Nullable final String state,
+            @NonNull final String email,
+            @Nullable final String userId,
+            @NonNull final String projectId
+    ) {
+        final PaymentCredentials pc = new PaymentCredentials();
         pc.generateId();
         if (brand == Brand.MASTERCARD || brand == Brand.AMEX || brand == Brand.VISA) {
             pc.type = Type.PAYONE_CREDITCARD;
@@ -490,22 +484,32 @@ public class PaymentCredentials {
         }
         pc.projectId = projectId;
 
-        List<X509Certificate> certificates = Snabble.getInstance().getPaymentCertificates();
-        if (certificates.size() == 0) {
+        final List<X509Certificate> certificates = Snabble.getInstance().getPaymentCertificates();
+        if (certificates == null || certificates.isEmpty()) {
             return null;
         }
 
-        PayoneData payoneData = new PayoneData(pseudocardpan, lastname, userId);
+        final PayoneData payoneData = new PayoneData(
+                pseudoCardPan,
+                lastname,
+                email,
+                street,
+                zip,
+                city,
+                country,
+                state,
+                userId
+        );
 
-        String json = GsonHolder.get().toJson(payoneData, PayoneData.class);
+        final String json = GsonHolder.get().toJson(payoneData, PayoneData.class);
 
-        X509Certificate certificate = certificates.get(0);
+        final X509Certificate certificate = certificates.get(0);
         pc.rsaEncryptedData = pc.rsaEncrypt(certificate, json.getBytes());
         pc.signature = pc.sha256Signature(certificate);
         pc.appId = Snabble.getInstance().getConfig().appId;
         pc.brand = brand;
-        pc.obfuscatedId = truncatedcardpan;
-        pc.validTo = parseValidTo("yyMM", cardexpiredate);
+        pc.obfuscatedId = truncatedCardPan;
+        pc.validTo = parseValidTo("yyMM", cardExpiryDate);
 
         if (pc.rsaEncryptedData == null) {
             return null;